Move template stuff in mail_template and use template.Must
